コードの分割、フォルダ構成整理
