+++ go.mod
@@ -1,4 +1,4 @@
-module main
+module GO-api
 
 go 1.24.2
