+++ internal/handler/hello.go
@@ -1,0 +1,14 @@
+package handler 
+
+import (
+	"fmt"
+	"net/http"
+)
+
+func HelloHandler(w http.ResponseWriter, r *http.Request) {
+	if r.Method != http.MethodGet {
+	  http.Error(w, "Method not allowed", http.StatusMethodNotAllowed)
+	  return
+	}
+	fmt.Fprintln(w, "Hello go!")
+  }
